DataFlow/requirements.txt
apache-beam[gcp]
pandas
oauth2client==3.0.0 
google-cloud-storage
scikit-learn
pickle

